Improved SciPlot used by the cartesian plot widget. This should be used in place of the XrtGraph code. The XrtGraph code is not open source and from a company that no longer exists. Some of the SciPlot changes include: 1. Ability to support multiple Y axes. When using multiple axes, the associated line colors are displayed around the axis label. 2. Added the step plot type and fill-under plot type. 3. Added time plot type for scalar Y values. 4. Plots update each second even if PVs don't update. Fixed a few Segfaults. Fixed issues with the color palette and the attributes screen.

+++ medm/medmSciPlot.h
@@ -23,8 +23,8 @@ typedef struct {
     int npoints;
     int pointsUsed;
     int listid;
-    float *xp;
-    float *yp;
+    double *xp;
+    double *yp;
 } CpDataSet;
 
 typedef struct {
